+++ src/main/java/com/ocp/ocp_finalproject/workflow/dto/response/WorkflowResponse.java
@@ -1,5 +1,6 @@
-package com.ocp.ocp_finalproject.workflow.dto;
+package com.ocp.ocp_finalproject.workflow.dto.response;
 
+import com.ocp.ocp_finalproject.workflow.dto.SetTrendCategoryDto;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -7,7 +8,7 @@ import lombok.Getter;
 @Getter
 @Builder
 @AllArgsConstructor
-public class WorkflowEditResponse {
+public class WorkflowResponse {
 
     private Long workflowId;
 
@@ -15,7 +16,7 @@ public class WorkflowEditResponse {
 
     private String siteUrl;
 
-    private Long blogTypeId;
+    private String blogType;
 
     private String blogUrl;
 
@@ -23,5 +24,6 @@ public class WorkflowEditResponse {
 
     private String blogAccountId;
 
-    private RecurrenceRuleDto recurrenceRule;
+    private String readableRule;
+
 }
